+++ classes/Pronamic/WP/Pay/Admin.php
@@ -13,6 +13,8 @@ class Pronamic_WP_Pay_Admin {
 	 * Constructs and initalize an admin object
 	 */
 	public function __construct() {
+		$this->settings = new Pronamic_WP_Pay_Settings();
+
 		// Actions
 		add_action( 'admin_init', array( $this, 'admin_init' ) );
 		add_action( 'admin_menu', array( $this, 'admin_menu' ) );
@@ -37,8 +39,6 @@ class Pronamic_WP_Pay_Admin {
 		$this->maybe_download_private_certificate();
 		$this->maybe_download_private_key();
 
-		$this->settings_init();
-
 		// Actions
 		// Show license message if the license is not valid
 		add_action( 'admin_notices', array( $this, 'admin_notices' ) );
@@ -46,7 +46,6 @@ class Pronamic_WP_Pay_Admin {
 		// Post types
 		new Pronamic_WP_Pay_Admin_GatewayPostType();
 		new Pronamic_WP_Pay_Admin_PaymentPostType();
-		new Pronamic_WP_Pay_Admin_PaymentFormPostType();
 
 		// Maybe update
 		global $pronamic_pay_db_version;
@@ -59,67 +58,6 @@ class Pronamic_WP_Pay_Admin {
 	}
 
 	//////////////////////////////////////////////////
-
-	public function settings_init() {
-		// Settings - General
-		add_settings_section(
-			'pronamic_pay_general', // id
-			__( 'General', 'pronamic_ideal' ), // title
-			array( __CLASS__, 'settings_section' ), // callback
-			'pronamic_pay' // page
-		);
-
-		add_settings_field(
-			'pronamic_pay_license_key', // id
-			__( 'Support License Key', 'pronamic_ideal' ), // title
-			array( __CLASS__, 'input_element' ), // callback
-			'pronamic_pay', // page
-			'pronamic_pay_general', // section
-			array( 'type' => 'password', 'label_for' => 'pronamic_pay_license_key' ) // args
-		);
-
-		register_setting( 'pronamic_pay', 'pronamic_pay_license_key', 'trim' );
-
-		// Settings - Pages
-		add_settings_section(
-			'pronamic_pay_pages', // id
-			__( 'Payment Status Pages', 'pronamic_ideal' ), // title
-			array( __CLASS__, 'settings_section' ), // callback
-			'pronamic_pay' // page
-		);
-
-		$pages = array(
-			'error'     => __( 'Error', 'pronamic_ideal' ),
-			'cancel'    => __( 'Canceled', 'pronamic_ideal' ),
-			'unknown'   => __( 'Unknown', 'pronamic_ideal' ),
-			'expired'   => __( 'Expired', 'pronamic_ideal' ),
-			'completed' => __( 'Completed', 'pronamic_ideal' ),
-		);
-
-		foreach ( $pages as $key => $label ) {
-			$id = sprintf( 'pronamic_pay_%s_page_id', $key );
-
-			add_settings_field(
-				$id, // id
-				$label, // title
-				array( __CLASS__, 'input_page' ), // callback
-				'pronamic_pay', // page
-				'pronamic_pay_pages', // section
-				array( 'label_for' => $id ) // args
-			);
-
-			register_setting( 'pronamic_pay', $id );
-		}
-	}
-
-	//////////////////////////////////////////////////
-
-	/**
-	 * Settings section
-	 */
-	public static function settings_section() {
-
-	}
 
 	public static function input_checkbox( $args ) {
 		$defaults = array(
@@ -211,49 +149,6 @@ class Pronamic_WP_Pay_Admin {
 		} else {
 			return $output;
 		}
-	}
-
-	/**
-	 * Input text
-	 *
-	 * @param array $args
-	 */
-	public static function input_element( $args ) {
-		$defaults = array(
-			'type' => 'text',
-		);
-
-		$args = wp_parse_args( $args, $defaults );
-
-		printf(
-			'<input name="%s" id="%s" type="%s" value="%s" class="%s" />',
-			esc_attr( $args['label_for'] ),
-			esc_attr( $args['label_for'] ),
-			esc_attr( $args['type'] ),
-			esc_attr( get_option( $args['label_for'] ) ),
-			'regular-text'
-		);
-
-		$status = get_option( 'pronamic_pay_license_status' );
-
-		$icon = 'valid' == $status ? 'yes' : 'no';
-
-		printf( '<span class="dashicons dashicons-%s" style="vertical-align: text-bottom;"></span>', esc_attr( $icon ) );
-	}
-
-	/**
-	 * Input page
-	 *
-	 * @param array $args
-	 */
-	public static function input_page( $args ) {
-		$name = $args['label_for'];
-
-		wp_dropdown_pages( array(
-			'name' => $name,
-			'selected' => get_option( $name, '' ),
-			'show_option_none' => __( '&mdash; Select a page &mdash;', 'pronamic_ideal' )
-		) );
 	}
 
 	//////////////////////////////////////////////////
@@ -382,7 +277,10 @@ class Pronamic_WP_Pay_Admin {
 			$gateway = Pronamic_WP_Pay_Plugin::get_gateway( $id );
 
 			if ( $gateway ) {
-				$amount = filter_input( INPUT_POST, 'test_amount', FILTER_VALIDATE_FLOAT );
+				$amount = filter_input( INPUT_POST, 'test_amount', FILTER_VALIDATE_FLOAT, array(
+					'flags'   => FILTER_FLAG_ALLOW_THOUSAND,
+					'options' => array( 'decimal' => pronamic_pay_get_decimal_separator() ),
+				) );
 
 				$data = new Pronamic_WP_Pay_PaymentTestData( wp_get_current_user(), $amount );