+++ classes/Pronamic/WP/Pay/Plugin.php
@@ -48,7 +48,7 @@ class Pronamic_WP_Pay_Plugin {
 		Pronamic_WPMUDEV_Membership_IDeal_AddOn::bootstrap();
 		// Pronamic_EShop_IDeal_AddOn::bootstrap();
 		Pronamic_WP_Pay_Extensions_EDD_Extension::bootstrap();
-		Pronamic_IThemesExchange_IDeal_AddOn::bootstrap();
+		Pronamic_WP_Pay_Extensions_IThemesExchange_Extension::bootstrap();
 
 		// Admin
 		if ( is_admin() ) {
@@ -69,7 +69,6 @@ class Pronamic_WP_Pay_Plugin {
 		require_once self::$dirname . '/includes/functions.php';
 		require_once self::$dirname . '/includes/formatting.php';
 		require_once self::$dirname . '/includes/page-functions.php';
-		require_once self::$dirname . '/includes/gravityforms.php';
 		require_once self::$dirname . '/includes/providers.php';
 		require_once self::$dirname . '/includes/gateways.php';
 		require_once self::$dirname . '/includes/payment.php';